[ci-skip] refactor npm scripts using gulp-shelter

+++ src/models/product.js
@@ -10,7 +10,7 @@ module.exports = (sequelize, DataTypes) => {
       required: true,
     },
     price: {
-      type:                     DataTypes.FLOAT,
+      type:                     DataTypes.INTEGER,
       required: false,
     },
   });
